Get rid of 800 bytes stack object
The last 800 bytes of firmware are checked for version info and this is
used to confirm the firmware "looks valid". This is a bit awkward for fw
compression as the code has to be extracted somewhere.

Load all but the last 800+511 bytes, then load the reset into
spid_buf that can be used to search for the version without having to
use another buffer.