Some reorg for loading firmware from flash
Rename _size to _len

Get rid of cyw43_firmware_details. This is now just another function
pointer. If CYW43_FIRMWARE_FLASH_BLOCK is defined we read the firmware
details from this page in flash.

cyw43_get_firmware_funcs gives you the function pointers, it's just a
macro that could be changed. We set it to functions that read compressed
or uncompressed data depending on CYW43_ENABLE_FIRMWARE_COMPRESSION