src: Add auth timeout error.
With some access points, if you don't disconnect properly, the next time
you connect the firmware reports an auth timeout which is converted by
the driver to CYW43_LINK_BADAUTH.

However if the auth timeout is ignored, the connection succeeds a split
second later. There's already code in the driver to handle "good-auth
follows a bad-auth", but this is happening too late.

This change adds a CYW43_LINK_TOAUTH error which can be ignored by the
client.